--- lib/context/index.js
@@ -1,64 +1,0 @@
-'use strict';
-
-const debug = require('debug')('graphql-component:context');
-
-const createContext = function (root, ctxConfig) {
-  return async function (context) {
-    const ctx = {};
-
-    for (const { component } of root.imports) {
-      Object.assign(ctx, await component._context(context));
-    }
-
-    if (ctxConfig) {
-      debug(`building ${ctxConfig.namespace} context`);
-      
-      if (!ctx[ctxConfig.namespace]) {
-        ctx[ctxConfig.namespace] = {};
-      }
-
-      Object.assign(ctx[ctxConfig.namespace], await ctxConfig.factory.call(root, context));
-    }
-
-    return ctx;
-  };
-};
-
-const wrapContext = function (component) {
-  const middleware = [];
-  const contextFunction = component._context;
-  const dataSourceInject = component._dataSourceInjection;
-
-  const context = async function (context) {
-    debug(`building ${component._id} root context`);
-
-    for (let { name, fn } of middleware) {
-      debug(`applying ${name} middleware`);
-      context = await fn(context);
-    }
-
-    const componentContext = await contextFunction(context);
-
-    const globalContext = {
-      ...context,
-      ...componentContext
-    };
-    
-    globalContext.dataSources = dataSourceInject(globalContext);
-    
-    return globalContext;
-  };
-
-  context.use = function (name, fn) {
-    if (typeof name === 'function') {
-      fn = name;
-      name = 'unknown';
-    }
-    debug(`adding ${name} middleware`);
-    middleware.push({ name, fn });
-  };
-
-  return context;
-};
-
-module.exports = { createContext, wrapContext };
